app/src/test/java/pacman/game/model/GameTest.java
/*
  VinUniversity
  Course: COMP1020 - Object-Oriented Programming & Data Structures
  Semester: Spring 2024
  Assessment: Term Project
  Author: Team 3
*/
package pacman.game.model;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class GameTest {
    private Game game;

    @BeforeEach
    public void setUp() {
        game = new Game();
    }

    @Test
    public void testGameInitialization() {
        assertNotNull(game.getGrid());
        assertNotNull(game.getPacMan());
        assertNotNull(game.getGhosts());
    }

    @Test
    public void testGameUpdate() {
        game.update();
        // Add assertions to check the game state after an update
    }
}
